+++ AdminLTE/app/Http/Livewire/ModuloCapacitaciones/Instructores/Index.php
@@ -4,6 +4,7 @@ namespace App\Http\Livewire\ModuloCapacitaciones\Instructores;
 
 use Livewire\Component;
 use App\Models\ModuloCapacitaciones\Instructore;
+use App\Models\ModuloCapacitaciones\Curso;
 use Livewire\WithPagination;
 
 class Index extends Component
@@ -13,14 +14,16 @@ class Index extends Component
     public $search = '';
     public $sort = 'id';
     public $direction = 'desc';
-    public $cant = '10';
+    public $cant = '5';
 
     protected $queryString = [
-        'cant' => ['except' => '10'], 
+        'cant' => ['except' => '5'], 
         'sort' => ['except' => 'id'], 
         'direction' => ['except' => 'desc'], 
         'search' => ['except' => '']
     ];
+
+    protected $paginationTheme = "bootstrap";
 
     protected $listeners = ['delete'];
     
@@ -34,15 +37,9 @@ class Index extends Component
     {
         $instructores = Instructore::select('instructores.id', 'instructores.resena', 'users.name', 'instructores.status')
                 ->join('users', 'instructores.user_id', '=', 'users.id')
-                ->where('instructores.resena', 'like', '%'. $this->search. '%')
-                ->orWhere('users.name', 'like', '%'. $this->search. '%')
+                ->where('users.name', 'like', '%'. $this->search. '%')
                 ->orderBy($this->sort, $this->direction)
                 ->paginate($this->cant);
-
-        // $instructores = Instructore::where('id', 'like' , '%' . $this->search . '%')
-        //             ->orWhere('resena', 'like' , '%' . $this->search . '%')
-        //             ->orderBy($this->sort, $this->direction)
-        //             ->paginate($this->cant);
 
         return view('livewire.modulo-capacitaciones.instructores.index', compact('instructores'));
 
@@ -65,6 +62,15 @@ class Index extends Component
 
     public function delete(Instructore $instructore)
     {
-        $instructore->delete();
+        $cursos = Curso::where('instructore_id', '=', $instructore->id)->get();
+        $contador = count($cursos);
+
+        if($contador > 0){
+            $this->emit('error', 'Este instructor no se puede eliminar, tiene cursos');
+        }else{
+            $instructore->delete();
+            $this->emit('alert', '¡Instructor eliminado con exito!');
+        }
+        
     }
 }
